# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CineSentiment API
  version: 2.0.0
  description: |
    REST API for IMDB movie-review binary sentiment classification using a fine-tuned
    DistilBERT model. v2 adds `/api/analyze`, ML aspect tagging, multilingual detection,
    SSE batch streaming, Developer API keys, Stripe billing, and outbound webhooks.
    Includes explainability, evaluation metrics, bootstrap CIs, hypothesis tests, and
    capstone readiness endpoints for coursework.
  contact:
    name: CineSentiment
  license:
    name: MIT
servers:
  - url: http://127.0.0.1:8000
    description: Local development server
  - url: http://localhost:8000
    description: Local development (alternate host)
...
paths:
...
  /api/analyze:
    post:
      tags: [Prediction]
      summary: Full analysis (predict + explain + sentence arc)
      description: |
        One-shot analysis for the web app: sentiment label, optional input×gradient explanation,
        and per-sentence tone arc when the review has 2+ sentences.
      operationId: analyzeReview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                explain:
                  type: boolean
                  default: true
                arc:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Analysis result with optional explanation and arc
        '400':
          description: Invalid input
...
  /api/predict:
    post:
      tags: [Prediction]
      summary: Predict sentiment
      description: |
        Classify sentiment for a single review (JSON) or a batch of reviews (multipart CSV).
        CSV uploads must include a `text` column. Returns a single `PredictionResult` for JSON
        or a `results` array for batch CSV.
      operationId: predictSentiment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                  description: Movie review text (truncated to server max length).
                  example: This movie was absolutely fantastic! Amazing acting and plot.
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with a required `text` column.
      responses:
        '200':
          description: Prediction result(s)
          content:
            application/json:
              schema:
                oneOf:
                  - allOf:
                      - $ref: '#/components/schemas/PredictionResult'
                      - type: object
                        properties:
                          model_source:
                            type: string
                            example: imdb_finetuned_local
                  - $ref: '#/components/schemas/BatchPredictionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
...
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingText:
              value:
                error: Text is required
            invalidSplit:
              value:
                error: split must be "val" or "test"
            invalidThreshold:
              value:
                error: Threshold must be between 0.0 and 1.0
    Forbidden:
      description: Cross-origin request blocked (CSRF origin check)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Cross-origin request blocked
...
    ServiceUnavailable:
      description: Model or evaluation artifact not ready
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            modelNotReady:
              value:
                error: Model not ready for inference
                detail: Train the model (scripts/model_training.py) or set HUB_MODEL_FALLBACK.
                model_source: untrained_base
            noArtifact:
              value:
                error: No evaluation artifact found.
                hint: "Run: make evaluate  (or make train-fast && make evaluate)"
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        detail:
          type: string
        hint:
          type: string
        model_source:
          type: string
        split:
          type: string
        path:
          type: string
...
    PredictionResult:
      type: object
      required:
        - label
        - probability
        - probability_positive
        - probability_negative
        - confidence
        - sentiment
      properties:
        label:
          type: integer
          enum: [0, 1]
          description: Predicted class (0 = negative, 1 = positive).
        probability:
          type: number
          format: float
          description: P(positive); same as probability_positive.
        probability_positive:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        probability_negative:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
        confidence:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          description: Max class probability for the predicted label.
        sentiment:
          type: string
          enum: [positive, negative]
        model_source:
          type: string
        text:
          type: string
          description: Present on batch CSV prediction rows.
    BatchPredictionResponse:
      type: object
      required: [results, model_source]
      properties:
        results:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/PredictionResult'
              - type: object
                required: [text]
        model_source:
          type: string
...